+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 ## [1.0.2] - 2022-04-04
 
 - first commit
+
+## February 28, 2023
+- add create ticket command